+++ tourcount/src/main/java/com/wmstein/tourcount/SettingsFragment.kt
@@ -4,6 +4,7 @@ import android.content.SharedPreferences
 import android.os.Bundle
 import androidx.preference.ListPreference
 import androidx.preference.PreferenceFragmentCompat
+import androidx.preference.SwitchPreferenceCompat
 import com.wmstein.tourcount.TourCountApplication.Companion.getPrefs
 
 /**
@@ -11,7 +12,7 @@ import com.wmstein.tourcount.TourCountApplication.Companion.getPrefs
  * Created by wmstein on 2020-04-17
  * last edited in Java on 2020-04-18,
  * converted to Kotlin on 2023-07-06,
- * last edited on 2025-11-01
+ * last edited on 2025-12-31
  */
 class SettingsFragment : PreferenceFragmentCompat() {
     private var prefs: SharedPreferences? = null
@@ -27,6 +28,11 @@ class SettingsFragment : PreferenceFragmentCompat() {
         val prefProx: Boolean = prefs!!.getBoolean("enable_prox", false)
         val proxPref: ListPreference? = findPreference("pref_prox")
         proxPref?.isEnabled = prefProx
+
+        // Set vibrator option visible if available in device
+        val prefVib: Boolean = prefs!!.getBoolean("", false)
+        val vibPref: SwitchPreferenceCompat? = findPreference("pref_button_vib")
+        vibPref?.isEnabled = prefVib
     }
 
 }