+++ tourcount/src/main/java/com/wmstein/tourcount/WelcomeActivity.java
@@ -24,6 +24,7 @@ import android.os.Bundle;
 import android.os.Environment;
 import android.os.Handler;
 import android.os.Looper;
+import android.os.Vibrator;
 import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
@@ -86,7 +87,7 @@ import java.util.Objects;
  * <p>
  * Based on BeeCount's WelcomeActivity.java by milo on 05/05/2014.
  * Changes and additions for TourCount by wmstein since 2016-04-18,
- * last edited on 2025-12-28
+ * last edited on 2025-12-31
  */
 public class WelcomeActivity
         extends AppCompatActivity
@@ -137,7 +138,7 @@ public class WelcomeActivity
         super.onCreate(savedInstanceState);
 
         if (IsRunningOnEmulator.DLOG || BuildConfig.DEBUG)
-            Log.i(TAG, "140, onCreate");
+            Log.i(TAG, "141, onCreate");
 
         tourCount = (TourCountApplication) getApplication();
 
@@ -153,9 +154,14 @@ public class WelcomeActivity
         // Grey out preferences menu item pref_prox when max. proximity sensitivity = null
         boolean prefProx = mProximity != null;
 
-        // Set pref_prox enabler, used in SettingsFragment
+        // Grey out preferences menu item pref_button_vib when device has no vibrator
+        Vibrator vibrator = getApplicationContext().getSystemService(Vibrator.class);
+        boolean prefVib = vibrator.hasVibrator();
+
+        // Set pref_prox and pref_button_vib enabler, used in SettingsFragment
         editor = prefs.edit();
         editor.putBoolean("enable_prox", prefProx);
+        editor.putBoolean("enable_vib", prefVib);
         editor.apply();
 
         // Set DarkMode when system is in BrightMode
@@ -202,7 +208,7 @@ public class WelcomeActivity
             editor.commit();
         }
         if (IsRunningOnEmulator.DLOG || BuildConfig.DEBUG)
-            Log.d(TAG, "205, onCreate, storagePermGranted: " + storagePermGranted);
+            Log.d(TAG, "211, onCreate, storagePermGranted: " + storagePermGranted);
 
         // Check DB version and upgrade if necessary
         dbHelper = new DbHelper(this);
@@ -315,7 +321,7 @@ public class WelcomeActivity
         super.onResume();
 
         if (IsRunningOnEmulator.DLOG || BuildConfig.DEBUG)
-            Log.i(TAG, "318, onResume");
+            Log.i(TAG, "324, onResume");
 
         prefs = TourCountApplication.getPrefs();
         prefs.registerOnSharedPreferenceChangeListener(this);
@@ -377,7 +383,7 @@ public class WelcomeActivity
         // Get new location self permission state
         isFineLocationPermGranted(); // set fineLocationPermGranted from self permission
         if (IsRunningOnEmulator.DLOG || BuildConfig.DEBUG)
-            Log.i(TAG, "380, onResume, fineLocationPermGranted: " + fineLocationPermGranted);
+            Log.i(TAG, "386, onResume, fineLocationPermGranted: " + fineLocationPermGranted);
 
         locationDispatcher(1);
     }
@@ -390,12 +396,12 @@ public class WelcomeActivity
             // check permission MANAGE_EXTERNAL_STORAGE for Android >= 11
             storageGranted = Environment.isExternalStorageManager();
             if (IsRunningOnEmulator.DLOG || BuildConfig.DEBUG)
-                Log.i(TAG, "393, ManageStoragePermission: " + storagePermGranted);
+                Log.i(TAG, "399, ManageStoragePermission: " + storagePermGranted);
         } else {
             storageGranted = ContextCompat.checkSelfPermission(this,
                     Manifest.permission.WRITE_EXTERNAL_STORAGE) == PackageManager.PERMISSION_GRANTED;
             if (IsRunningOnEmulator.DLOG || BuildConfig.DEBUG)
-                Log.i(TAG, "398, ExtStoragePermission: " + storagePermGranted);
+                Log.i(TAG, "404, ExtStoragePermission: " + storagePermGranted);
         }
         return storageGranted;
     }
@@ -420,7 +426,7 @@ public class WelcomeActivity
                     // Get location data
                     if (!locServiceOn) {
                         if (IsRunningOnEmulator.DLOG || BuildConfig.DEBUG)
-                            Log.i(TAG, "423, locationDispatcher 1");
+                            Log.i(TAG, "429, locationDispatcher 1");
                         Intent sIntent = new Intent(getApplicationContext(), LocationService.class);
                         startService(sIntent);
                         locServiceOn = true;
@@ -434,7 +440,7 @@ public class WelcomeActivity
                     // Stop location service
                     if (locServiceOn) {
                         if (IsRunningOnEmulator.DLOG || BuildConfig.DEBUG)
-                            Log.i(TAG, "437, locationDispatcher 2");
+                            Log.i(TAG, "443, locationDispatcher 2");
                         locationService.stopListener();
                         Intent sIntent = new Intent(getApplicationContext(), LocationService.class);
                         stopService(sIntent);
@@ -639,7 +645,7 @@ public class WelcomeActivity
         super.onPause();
 
         if (IsRunningOnEmulator.DLOG || BuildConfig.DEBUG)
-            Log.i(TAG, "642, onPause");
+            Log.i(TAG, "648, onPause");
 
         headDataSource.close();
         countDataSource.close();
@@ -655,7 +661,7 @@ public class WelcomeActivity
         super.onStop();
 
         if (IsRunningOnEmulator.DLOG || BuildConfig.DEBUG)
-            Log.i(TAG, "658 onStop");
+            Log.i(TAG, "664 onStop");
 
         baseLayout.invalidate();
     }
@@ -665,7 +671,7 @@ public class WelcomeActivity
         super.onDestroy();
 
         if (IsRunningOnEmulator.DLOG || BuildConfig.DEBUG)
-            Log.i(TAG, "668, onDestroy");
+            Log.i(TAG, "674, onDestroy");
     }
 
     // Handle button click "Counting" here
@@ -713,7 +719,7 @@ public class WelcomeActivity
     // Import the basic DB
     private void importBasisDb() {
         if (IsRunningOnEmulator.DLOG || BuildConfig.DEBUG)
-            Log.d(TAG, "716, importBasicDBFile");
+            Log.d(TAG, "722, importBasicDBFile");
 
         String fileExtension = ".db";
         String fileNameStart = "tourcount0";
@@ -758,7 +764,7 @@ public class WelcomeActivity
                         if (data != null) {
                             selectedFile = data.getStringExtra("fileSelected");
                             if (IsRunningOnEmulator.DLOG || BuildConfig.DEBUG)
-                                Log.i(TAG, "761, Selected file: " + selectedFile);
+                                Log.i(TAG, "767, Selected file: " + selectedFile);
 
                             if (selectedFile != null)
                                 inFile = new File(selectedFile);
@@ -862,7 +868,7 @@ public class WelcomeActivity
                         if (data != null) {
                             selectedFile = data.getStringExtra("fileSelected");
                             if (IsRunningOnEmulator.DLOG || BuildConfig.DEBUG)
-                                Log.d(TAG, "865, File selected: " + selectedFile);
+                                Log.d(TAG, "871, File selected: " + selectedFile);
 
                             if (selectedFile != null)
                                 inFile = new File(selectedFile);
@@ -1025,7 +1031,7 @@ public class WelcomeActivity
                 }
             } catch (IOException e) {
                 if (IsRunningOnEmulator.DLOG || BuildConfig.DEBUG)
-                    Log.e(TAG, "1028, Failed to export Basic DB");
+                    Log.e(TAG, "1034, Failed to export Basic DB");
                 mesg = getString(R.string.saveFail);
                 Toast.makeText(this,
                         HtmlCompat.fromHtml("<font color='red'><b>" + mesg + "</b></font>",
@@ -1587,7 +1593,7 @@ public class WelcomeActivity
                     if (longi != 0) // Has coordinates
                     {
                         if (IsRunningOnEmulator.DLOG || BuildConfig.DEBUG)
-                            Log.d(TAG, "1590 longi " + longi);
+                            Log.d(TAG, "1596 longi " + longi);
                         if (frst == 0) {
                             loMin = longi;
                             loMax = longi;
@@ -1667,7 +1673,7 @@ public class WelcomeActivity
                         HtmlCompat.fromHtml("<font color='red'><b>" + mesg + "</b></font>",
                                 HtmlCompat.FROM_HTML_MODE_LEGACY), Toast.LENGTH_LONG).show();
                 if (IsRunningOnEmulator.DLOG || BuildConfig.DEBUG)
-                    Log.e(TAG, "1670, Failed to export csv file");
+                    Log.e(TAG, "1676, Failed to export csv file");
             }
         }
     }
@@ -1875,7 +1881,7 @@ public class WelcomeActivity
             dbHelper.close();
         } catch (Exception e) {
             if (IsRunningOnEmulator.DLOG || BuildConfig.DEBUG)
-                Log.e(TAG, "1878, Failed to reset DB");
+                Log.e(TAG, "1884, Failed to reset DB");
 
             mesg = getString(R.string.resetFail);
             Toast.makeText(this,